Added blueprint for posts route, deleted all exisiting data of a user upon deletion
